Adding match information to the match table
Added match information to the match table
for each othe players listed in the summoner
table. The value insertion works fine but
since this pull of data is the largest so far
we're hitting the API timeout limits.

Right now there is an issue that even after a
full 2 minute timeout from the API a JSON object
yelling at us about timeouts gets passed on instead
of a MatchDto.

I knew that the timeout solution I wrote was pretty janky,
but didn't realize how janky so next step is to look for
a better timeout handler.