Build Match revised for less redundant API calls
buildMatch revised so that based on one call from the match api all
players involved are added to the match. So when we are querying for
Player A to get their match details, we see if any other players
in the database were also in that game. Previously if multiple players
were in a game, the match details would get queried multiple times.