buildMatchHistory has been changed to updateMatchHistory
buildMatchHistory's implementation has been slightly changed so that is
used both in initial creation of the match table, and further updates to
the match table.

The timestamp of the most recent game in the table (0 if the table is
empty) is compared against the timestamp of games in the queried
matchlist. If we reach a game from the matchList that is less recent
then that matchList is discarded and we begin querying for the matchList
of the next summoner.